requirements.txt
fastmcp
python-jose[cryptography]
httpx
python-dotenv